Use ajax to fetch app count
Allow browser to cache it for 3600 seconds.

Do not use server side cache for now.
